Added the force option to CLI and alter all classes and tests to facilitate it.

+++ lib/dbgeni/cli.rb
@@ -3,6 +3,7 @@ $:.unshift File.expand_path(File.join(File.dirname(__FILE__), "..", "..", "lib")
 
 $config_file      = './.dbgeni'
 $environment_name = nil
+$force            = false
 
 if index = ARGV.index('--config-file') or index = ARGV.index('-c')
   unless ARGV[index+1]
@@ -23,6 +24,11 @@ if index = ARGV.index('--environment-name') or index = ARGV.index('-e')
   $environment_name = ARGV[index+1]
   # remove all references to config file from the argument list
   ARGV.delete_at(index)
+  ARGV.delete_at(index)
+end
+
+if index = ARGV.index('--force') or index = ARGV.index('-f')
+  $force = true
   ARGV.delete_at(index)
 end
